@@ -1,4 +1,3 @@
-
 import 'package:flutter/material.dart';
 import 'dart:math' as math;
 
@@ -18,14 +17,14 @@ class IconToggle extends StatefulWidget {
     this.onChanged,
     this.transitionBuilder = _defaultTransitionBuilder,
     this.duration = const Duration(milliseconds: 100),
-    this.reverseDuration,
+    required this.reverseDuration,
   });
   final IconData selectedIconData;
   final IconData unselectedIconData;
   final Color activeColor;
   final Color inactiveColor;
   final bool value;
-  final ValueChanged<bool> onChanged;
+  final ValueChanged<bool>? onChanged;
   final AnimatedSwitcherTransitionBuilder transitionBuilder;
   final Duration duration;
   final Duration reverseDuration;
@@ -35,30 +34,31 @@ class IconToggle extends StatefulWidget {
 
 class _IconToggleState extends State<IconToggle>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  Animation<double> _position;
+  late final AnimationController _controller;
+  late final Animation<double> _position;
   bool _cancel = false;
 
   @override
   void initState() {
     super.initState();
     _controller = AnimationController(
-        vsync: this,
-        duration: Duration(milliseconds: 100),
-        reverseDuration: Duration(milliseconds: 50));
+      vsync: this,
+      duration: Duration(milliseconds: 100),
+      reverseDuration: Duration(milliseconds: 50),
+    );
     _position = CurvedAnimation(parent: _controller, curve: Curves.linear);
     _position.addStatusListener((status) {
       if (status == AnimationStatus.dismissed &&
           widget.onChanged != null &&
           _cancel == false) {
-        widget.onChanged(!widget.value);
+        widget.onChanged!(!widget.value);
       }
     });
   }
 
   @override
   void dispose() {
-    _controller?.dispose();
+    _controller.dispose();
     super.dispose();
   }
 
@@ -68,14 +68,14 @@ class _IconToggleState extends State<IconToggle>
       behavior: HitTestBehavior.opaque,
       onTapDown: (event) {
         _cancel = false;
-        _controller?.forward();
+        _controller.forward();
       },
       onTapUp: (event) {
-        _controller?.reverse();
+        _controller.reverse();
       },
       onTapCancel: () {
         _cancel = true;
-        _controller?.reverse();
+        _controller.reverse();
       },
       child: Padding(
         padding: const EdgeInsets.all(10.0),
@@ -88,7 +88,9 @@ class _IconToggleState extends State<IconToggle>
             reverseDuration: widget.reverseDuration,
             transitionBuilder: widget.transitionBuilder,
             child: Icon(
-              widget.value ? widget.selectedIconData : widget.unselectedIconData,
+              widget.value
+                  ? widget.selectedIconData
+                  : widget.unselectedIconData,
               color: widget.value ? widget.activeColor : widget.inactiveColor,
               size: 22,
               key: ValueKey<bool>(widget.value),
@@ -102,14 +104,15 @@ class _IconToggleState extends State<IconToggle>
 
 class _IconToggleable<T> extends AnimatedWidget {
   _IconToggleable({
-    Animation<T> listenable,
-    this.activeColor,
-    this.inactiveColor,
-    this.child,
+    required this.listenable,
+    required this.activeColor,
+    required this.inactiveColor,
+    required this.child,
   }) : super(listenable: listenable);
   final Color activeColor;
   final Color inactiveColor;
   final Widget child;
+  final Animation<double> listenable;
   @override
   Widget build(BuildContext context) {
     return CustomPaint(
@@ -125,20 +128,20 @@ class _IconToggleable<T> extends AnimatedWidget {
 
 class _IconPainter extends CustomPainter {
   _IconPainter({
-    @required this.position,
-    this.activeColor,
-    this.inactiveColor,
+    required this.position,
+    required this.activeColor,
+    required this.inactiveColor,
   });
   final Animation<double> position;
   final Color activeColor;
   final Color inactiveColor;
 
-  double get _value => position != null ? position.value : 0;
+  double get _value => position.value;
 
   @override
   void paint(Canvas canvas, Size size) {
     final Paint paint = Paint()
-      ..color = Color.lerp(inactiveColor, activeColor, _value)
+      ..color = Color.lerp(inactiveColor, activeColor, _value)!
           .withOpacity(math.min(_value, 0.15))
       ..style = PaintingStyle.fill
       ..strokeWidth = 2.0;
